requirements.txt
scrapy<3
psycopg2-binary<3
python-dotenv<2